feat: WIP new places page [EP-2187]
