+++ packages/earth-admin/src/components/card/index.ts
@@ -1,0 +1,1 @@
+export { Card } from './Card';
